requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
numpy>=1.26.0
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0
groq==0.4.1
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.5

